+++ lib/maze/option.rb
@@ -14,11 +14,9 @@ module Maze
     NULL_PORT = 'null-port'
 
     # Appium options
-    SEPARATE_SESSIONS = 'separate-sessions'
     FARM = 'farm'
     APP = 'app'
     A11Y_LOCATOR = 'a11y-locator'
-    RESILIENT = 'resilient'
     CAPABILITIES = 'capabilities'
 
     # Generic device farm options