+++ test/hooks/appium_hooks_test.rb
@@ -6,7 +6,6 @@ require_relative '../../lib/maze/capabilities'
 require_relative '../../lib/maze/farm/browser_stack/capabilities'
 require_relative '../../lib/maze/wait'
 require_relative '../../lib/maze/driver/appium'
-require_relative '../../lib/maze/driver/resilient_appium'
 require_relative '../../lib/maze/hooks/appium_hooks'
 
 class AppiumHooksTest < Test::Unit::TestCase
@@ -63,27 +62,7 @@ class AppiumHooksTest < Test::Unit::TestCase
     assert_equal(:app, caps['app'])
   end
 
-  def test_create_driver_resilient
-    $config.expects(:resilient).returns(true)
-    $config.expects(:appium_server_url).returns(:appium_server_url)
-    $config.expects(:capabilities).returns(:capabilities)
-    $config.expects(:locator).returns(:locator)
-    $logger.expects(:info).with('Creating ResilientAppium driver instance')
-
-    Maze::Driver::ResilientAppium.expects(:new).with(
-      :appium_server_url,
-      :capabilities,
-      :locator
-    ).returns(:driver)
-
-    hooks = Maze::Hooks::AppiumHooks.new
-    driver = hooks.create_driver($config)
-
-    assert_equal(:driver, driver)
-  end
-
   def test_create_driver_default
-    $config.expects(:resilient).returns(false)
     $config.expects(:appium_server_url).returns(:appium_server_url)
     $config.expects(:capabilities).returns(:capabilities)
     $config.expects(:locator).returns(:locator)
@@ -113,7 +92,6 @@ class AppiumHooksTest < Test::Unit::TestCase
     hooks.expects(:create_driver).with($config).returns(driver_mock)
 
     $config.expects(:capabilities=).with(:caps)
-    $config.expects(:appium_session_isolation).returns(false)
     $config.expects(:farm).returns(:bs)
     $config.stubs(:device_list).returns([])
 
@@ -133,7 +111,6 @@ class AppiumHooksTest < Test::Unit::TestCase
     hooks.expects(:create_driver).with($config).returns(driver_mock)
 
     $config.expects(:capabilities=).with(:caps)
-    $config.expects(:appium_session_isolation).returns(false)
     $config.expects(:farm).returns(:local)
     $config.expects(:os_version).returns(nil)
     $config.expects(:os).returns('ios')
@@ -158,7 +135,6 @@ class AppiumHooksTest < Test::Unit::TestCase
     hooks.expects(:create_driver).with($config).returns(driver_mock)
 
     $config.expects(:capabilities=).with(:caps)
-    $config.expects(:appium_session_isolation).returns(false)
     $config.expects(:farm).returns(:local)
     $config.expects(:os_version).returns(nil)
     $config.expects(:os).returns('android')
@@ -181,7 +157,6 @@ class AppiumHooksTest < Test::Unit::TestCase
     hooks.expects(:create_driver).with($config).returns(driver_mock)
 
     $config.expects(:capabilities=).with(:caps)
-    $config.expects(:appium_session_isolation).returns(false)
     $config.expects(:device_list).returns([]).twice()
 
     $logger.expects(:error).with("Appium driver failed to start after 6 attempts in 60 seconds")
@@ -202,7 +177,6 @@ class AppiumHooksTest < Test::Unit::TestCase
     hooks.expects(:create_driver).with($config).returns(driver_mock)
 
     $config.expects(:capabilities=).with(:caps)
-    $config.expects(:appium_session_isolation).returns(false)
     $config.expects(:farm).returns(:farm)
     $config.expects(:device_list).returns([]).twice()
 
@@ -222,7 +196,6 @@ class AppiumHooksTest < Test::Unit::TestCase
     hooks.expects(:create_driver).twice.with($config).returns(driver_mock)
 
     $config.expects(:capabilities=).twice.with(:caps)
-    $config.expects(:appium_session_isolation).twice.returns(false)
     $config.expects(:device).twice.returns(:device)
     $config.expects(:farm).twice.returns(:farm)
     device_list = mock('device_list')
@@ -253,7 +226,6 @@ class AppiumHooksTest < Test::Unit::TestCase
     hooks.expects(:create_driver).times(3).with($config).returns(driver_mock)
 
     $config.expects(:capabilities=).times(3).with(:caps)
-    $config.expects(:appium_session_isolation).times(3).returns(false)
     $config.expects(:device).times(5).returns(:device)
     $config.expects(:farm).times(3).returns(:farm)
     device_list = mock('device_list')
@@ -273,23 +245,4 @@ class AppiumHooksTest < Test::Unit::TestCase
       hooks.start_driver($config)
     end
   end
-
-  def test_start_driver_session_isolation
-    driver_mock = mock('driver')
-
-    Maze.expects(:driver).twice.returns(false, true)
-    Maze.expects(:driver=).with(driver_mock)
-
-    hooks = Maze::Hooks::AppiumHooks.new
-    hooks.expects(:device_capabilities).with($config, nil).returns(:caps)
-    hooks.expects(:create_driver).with($config).returns(driver_mock)
-
-    $config.expects(:capabilities=).with(:caps)
-    $config.expects(:appium_session_isolation).returns(true)
-    $config.expects(:farm).returns(:farm)
-    $config.stubs(:device_list).returns([])
-
-    hooks.start_driver($config)
-  end
-
 end