+++ test/option/parser_test.rb
@@ -26,11 +26,9 @@ class ParserTest < Test::Unit::TestCase
     options = Maze::Option::Parser.parse args
 
     # Common options
-    assert_false(options[Maze::Option::SEPARATE_SESSIONS])
     assert_nil(options[Maze::Option::FARM])
     assert_nil(options[Maze::Option::APP])
     assert_false(options[Maze::Option::A11Y_LOCATOR])
-    assert_false(options[Maze::Option::RESILIENT])
     assert_equal('{}', options[Maze::Option::CAPABILITIES])
 
     # Device-farm-only options
@@ -65,11 +63,9 @@ class ParserTest < Test::Unit::TestCase
 
   def test_overwritten_values
     args = %w[
-      --separate-sessions
       --farm=ARG_FARM
       --app=ARG_APP
       --a11y-locator
-      --resilient
       --capabilities=ARG_CAPABILITIES
       --bs-local=ARG_BS_LOCAL
       --device=ARG_DEVICE
@@ -89,11 +85,9 @@ class ParserTest < Test::Unit::TestCase
     options = Maze::Option::Parser.parse args
 
     # Common options
-    assert_true(options[Maze::Option::SEPARATE_SESSIONS])
     assert_equal('ARG_FARM', options[Maze::Option::FARM])
     assert_equal('ARG_APP', options[Maze::Option::APP])
     assert_true(options[Maze::Option::A11Y_LOCATOR])
-    assert_true(options[Maze::Option::RESILIENT])
     assert_equal('ARG_CAPABILITIES', options[Maze::Option::CAPABILITIES])
 
     # Device-farm-only options